+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Use microcopy translations in buttons
 * Add `Action` class for user interaction
 * Add context-menu support to `alchemy-table`
+* Add support for using custom templates for `alchemy-select` options
 
 ## 0.1.5 (2022-03-21)